+++ src/testunits/module_BN/BayesNetTestSuite.h
@@ -747,7 +747,7 @@ namespace gum_tests {
 
       TS_ASSERT_EQUALS(i.domainSize(), (gum::Size)(2 * 2 * 2 * 2 * 3));
 
-      float a = 0.0;
+      double a = 0.0;
       TS_ASSERT_THROWS_NOTHING(a = bn.jointProbability(i));
 
       TS_ASSERT_DELTA(a, 0.00072, 1e-4);
@@ -757,7 +757,7 @@ namespace gum_tests {
         TS_ASSERT_THROWS(bn.jointProbability(j), gum::InvalidArgument);
         j.add(bn.variable(n));
       }
-      float b = 0.0;
+      double b = 0.0;
       TS_GUM_ASSERT_THROWS_NOTHING(b = bn.jointProbability(j));
 
       TS_ASSERT_EQUALS(a, b);
